requirements.txt
py==1.4.30
pytest==2.8.2
python-Levenshtein==0.12.0
wheel==0.24.0
